admins: updated get_course_rating_info and remove_student_from_course, crud_admin: updated: remove_student-from_course

+++ src/app/crud/crud_admin.py
@@ -2,15 +2,19 @@ from sqlalchemy.orm import Session
 from db.models import Student, StudentCourse, Course, Teacher, StudentRating, Account
 
 
-async def remove_student_from_course(db: Session, student_id: int, course_id) -> None:
+async def remove_student_from_course(db: Session, student_id: int, course_id) -> bool:
     query = (db.query(StudentCourse).filter(
         StudentCourse.student_id == student_id,
         StudentCourse.course_id == course_id
     ).first())
+    
+    if not query:
+        return False
 
     db.delete(query)
     db.commit()
-
+    
+    return True
 
 async def hide_course(db: Session, course: Course) -> None:
     course.is_hidden = True
